+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to **pipecat** will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.15] - 2024-05-15
+
+### Fixed
+
+- Quick hot fix for receiving `DailyTransportMessage`.
+
 ## [0.0.14] - 2024-05-15
 
 ### Added
@@ -15,7 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Images are now resized to the size of the output camera.
+- Images are now resized to the size of the output camera. This was causing
+  images not being displayed.
 
 - Fixed an issue in `DailyTransport` that would not allow the input processor to
   shutdown if no participant ever joined the room.